(feat) refactor fetch_trading_pairs to *_api_order_book.py file
